MetadataDocument moved to model
